+++ packages/e2e-tests/tests/common/atoms-tests/bid.ts
@@ -2,13 +2,14 @@ import type { IRaribleSdk } from "@rarible/sdk/src/domain"
 import type { BlockchainWallet } from "@rarible/sdk-wallet"
 import type { Order } from "@rarible/api-client"
 import type { OrderRequest, PrepareOrderRequest } from "@rarible/sdk/src/types/order/common"
+import type { CollectionId } from "@rarible/api-client"
 
 /**
  * Make new bid order
  */
 export async function bid(sdk: IRaribleSdk,
 						   wallet: BlockchainWallet,
-						   prepareOrderRequest: PrepareOrderRequest,
+						   prepareOrderRequest: PrepareOrderRequest | { collectionId: CollectionId },
 						   orderRequest: OrderRequest): Promise<Order> {
 	console.log("bid, prepare_order_update_request=", prepareOrderRequest)
 	const bidPrepare = await sdk.order.bid(prepareOrderRequest)