+++ packages/e2e-tests/tests/smoke/nft/mint-transfer-burn.test.ts
@@ -5,7 +5,7 @@ import type { MintRequest } from "@rarible/sdk/build/types/nft/mint/mint-request
 import type { BlockchainWallet } from "@rarible/sdk-wallet"
 import type { BurnRequest } from "@rarible/sdk/build/types/nft/burn/domain"
 import type { TransferRequest } from "@rarible/sdk/build/types/nft/transfer/domain"
-import { getEthereumWallet, getTezosTestWallet, getWalletAddressFull } from "../../common/wallet"
+import { getEthereumWallet, getSolanaWallet, getTezosTestWallet, getWalletAddressFull } from "../../common/wallet"
 import { createSdk } from "../../common/create-sdk"
 import { mint } from "../../common/atoms-tests/mint"
 import { testsConfig } from "../../common/config"
@@ -213,6 +213,37 @@ function suites(): {
 				creators: [],
 			},
 			totalBalanceAfterBurn: 16,
+		},
+		{
+			blockchain: Blockchain.SOLANA,
+			description: "NFT",
+			wallets: { creator: getSolanaWallet(0), recipient: getSolanaWallet(1) },
+			collectionId: testsConfig.variables.SOLANA_COLLECTION,
+			mintRequest: (walletAddress: UnionAddress): MintRequest => {
+				return {
+					uri: "https://arweave.net/Vt0uj2ql0ck-U5dLWDWJnwQaZPrvqkfxils8agrTiOc",
+					creators: [{
+						account: walletAddress,
+						value: 10000,
+					}],
+					royalties: [],
+					lazyMint: false,
+					supply: 1,
+				}
+			},
+			transferRequest: (walletAddress: UnionAddress): TransferRequest => {
+				return {
+					to: walletAddress,
+					amount: 1,
+				}
+			},
+			creatorBalanceAfterTransfer: "0",
+			recipientBalanceAfterTransfer: "1",
+			burnRequest: {
+				amount: 1,
+				creators: [],
+			},
+			totalBalanceAfterBurn: 0,
 		},
 	]
 	return allBlockchains.filter(b => testsConfig.blockchain?.includes(b.blockchain))