+++ packages/sdk/src/sdk-blockchains/solana/common/currencies.ts
@@ -1,0 +1,6 @@
+import { Blockchain } from "@rarible/api-client"
+import type { CurrencyType } from "../../../common/domain"
+
+export function getCurrencies():  CurrencyType[] {
+	return [{ blockchain: Blockchain.SOLANA, type: "NATIVE" }]
+}